Add UsuarioController and improve user service
Delete TesteController; add UsuarioController with /cadastrar and /login endpoints (CORS enabled). Change Pessoa.usuarioId and Usuario.usuarioId types from Long to Integer. Refactor UsuarioService: inject UsuarioRepository and PessoaRepository, add transactional cadastrarNovoUsuario to save Pessoa and create Usuario, and add realizarLogin for authentication. Adjusted responses and error handling for registration and login.

+++ src/main/java/com/feedbackplus/backend/model/Pessoa.java
@@ -13,7 +13,7 @@ public class Pessoa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "pessoa_id")
-    private Long pessoaId;
+    private Integer pessoaId;
 
     private String nome;
     private String cpf;
